Backend/requirements.txt
pdf2image
pytesseract
Pillow
fastapi[all]
opencv_python